ok, rather than do them all individually here's the lot.  this replaces the commit i just did.
there are 2 meshes in the drivers.blend file (root drivers folder) it has a driver1 with a visor and a driver2 with a visor.  if you want to get regular driver1 you just delete the visor.  driver1_visors's texture has an alpha layer with a partially opaque visor.

driver2's texture is all opaque.

these meshes look a little nicer around the hands and feet.  but the textures are still a bit weak.

also, if i'm doing anything poorly in this commit please tell me, not too used to svn type stuff.